- Sửa: tách factory thành file riêng khỏi ViewModel
