+++ app/src/main/java/com/datn/viettech_md_12/screen/post/PostScreen.kt
@@ -66,8 +66,8 @@ import com.datn.viettech_md_12.R
 import com.datn.viettech_md_12.component.MyTextField
 import com.datn.viettech_md_12.data.model.AllPostMetadata
 import com.datn.viettech_md_12.data.model.PostMetadata
+import com.datn.viettech_md_12.utils.PostViewModelFactory
 import com.datn.viettech_md_12.viewmodel.PostViewModel
-import com.datn.viettech_md_12.viewmodel.PostViewModelFactory
 import java.text.SimpleDateFormat
 import java.util.Date
 import java.util.Locale
