+++ app/src/main/java/com/datn/viettech_md_12/viewmodel/PostViewModel.kt
@@ -182,14 +182,3 @@ class PostViewModel(application: Application) : ViewModel() {
 
 }
 
-class PostViewModelFactory(
-    private val application: Application
-) : ViewModelProvider.AndroidViewModelFactory(application) {
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(PostViewModel::class.java)) {
-            @Suppress("UNCHECKED_CAST")
-            return PostViewModel(application) as T
-        }
-        throw IllegalArgumentException("Unknown ViewModel class")
-    }
-}
